[DF-14789] "Read only" actions and invisible actions. "mainwindow.ui" cleanup.
Former-commit-id: de19dcfe4c1706a2bfc7b4f916b2170a95111d82

+++ Modules/TArc/Sources/TArc/SharedModules/ActionManagmentModule/Private/ShortcutsModel.cpp
@@ -143,7 +143,7 @@ int ShortcutsModel::rowCount(const QModelIndex& parent) const
         QString blockName = blocks[parent.row()];
         auto iter = actions.find(blockName);
         DVASSERT(iter != actions.end());
-        return iter->second.size();
+        return static_cast<int>(iter->second.size());
     }
 
     return 0;
